feat: make LOADTEST mode the default emulated mode, rename simple mode to custom-emulated
Swaps mode names: multi-emulated/MULTI_EMULATED becomes emulated/EMULATED
(the new recommended default for large-scale tests), while the previous
single-process emulated/EMULATED mode becomes custom-emulated/CUSTOM_EMULATED.
Applies across the Java enum, browser-emulator TS types/validation, JSON
schema, OpenAPI docs, config examples, e2e test fixtures/scripts (renamed
accordingly), and README, and documents measured participants-per-worker
capacity on a t3.medium (baseline and with OpenVidu Pro/mediasoup).
