+++ browser-emulator/docs/openapi.yaml
@@ -168,7 +168,7 @@ paths:
       description: >
         Starts a single `lk load-test` process that simulates many
         publishers/subscribers in one room. Used by the controller when a test
-        case's `browser` is set to `multi-emulated`. Simulcast is enabled by
+        case's `browser` is set to `emulated`. Simulcast is enabled by
         default. Unlike the streamManager path, this does not publish
         custom/reference media (lk load-test uses its own synthetic clip).
       operationId: runLoadTest
@@ -451,7 +451,7 @@ components:
           enum:
             - chrome
             - firefox
-            - emulated
+            - custom-emulated
         recording:
           type: boolean
         showVideoElements:
@@ -467,7 +467,7 @@ components:
           description: >-
             Debug-only. For PUBLISHER participants, disables auto-subscription
             so the publisher doesn't also subscribe to other participants'
-            tracks. For emulated publishers it omits --auto-subscribe from the
+            tracks. For custom-emulated publishers it omits --auto-subscribe from the
             underlying `lk room join` command; for real-browser publishers it
             sets autoSubscribe=false in the LiveKit room connect options. Has no
             effect on SUBSCRIBER participants.
