+++ browser-emulator/package.json
@@ -35,12 +35,12 @@
 		"test:docker:e2e": "docker compose -f docker-compose.test.yml run --rm -e TEST_SCRIPT='test:e2e' --build browser-emulator-tests",
 		"test:docker:perf:benchmarks": "bash ./dev_scripts/run-perf-compare.sh benchmarks",
 		"test:docker:perf:saturation": "bash ./dev_scripts/run-perf-compare.sh saturation",
-		"vitest:e2e:emulated": "vitest run --project=e2e tests/e2e/emulated-browsers.test.ts",
-		"test:docker:emulated": "docker compose -f docker-compose.test.yml run --rm -e TEST_SCRIPT='vitest:e2e:emulated' --build browser-emulator-tests",
+		"vitest:e2e:custom-emulated": "vitest run --project=e2e tests/e2e/custom-emulated-browsers.test.ts",
+		"test:docker:custom-emulated": "docker compose -f docker-compose.test.yml run --rm -e TEST_SCRIPT='vitest:e2e:custom-emulated' --build browser-emulator-tests",
 		"vitest:e2e:real-browsers": "vitest run --project=e2e tests/e2e/real-browsers.test.ts",
 		"test:docker:real-browsers": "docker compose -f docker-compose.test.yml run --rm -e TEST_SCRIPT='vitest:e2e:real-browsers' --build browser-emulator-tests",
-		"vitest:e2e:multi-emulated": "vitest run --project=e2e tests/e2e/multi-emulated-browsers.test.ts",
-		"test:docker:multi-emulated": "docker compose -f docker-compose.test.yml run --rm -e TEST_SCRIPT='vitest:e2e:multi-emulated' --build browser-emulator-tests",
+		"vitest:e2e:emulated": "vitest run --project=e2e tests/e2e/emulated-browsers.test.ts",
+		"test:docker:emulated": "docker compose -f docker-compose.test.yml run --rm -e TEST_SCRIPT='vitest:e2e:emulated' --build browser-emulator-tests",
 		"qoe": "tsx ./src/qoe-cli.ts"
 	},
 	"author": "Carlos Santos & Iván Chicano",
