+++ browser-emulator/tests/e2e/custom-emulated-browsers.test.ts
@@ -1,0 +1,32 @@
+import { describe, it, beforeEach, afterEach } from 'vitest';
+import { startServer } from '../../src/app.js';
+import type { Application } from 'express';
+import {
+	run2BrowserTest,
+	cleanupServer,
+	setupServerPorts,
+} from './e2e-test-utils.js';
+
+let app: Application;
+
+beforeEach(async () => {
+	await setupServerPorts();
+}, 30000);
+
+afterEach(async () => {
+	await cleanupServer();
+}, 30000);
+
+// IMPORTANT: This test assumes there is a LiveKit server running and accessible with the credentials specified in test configs
+describe('Browser-emulator - Custom-emulated browsers', () => {
+	describe('LiveKit', () => {
+		beforeEach(async () => {
+			process.env.COM_MODULE = 'livekit';
+			({ app } = await startServer());
+		});
+		// Added repeats to these tests to increase confidence in stability, as browsers can be flaky
+		it('LiveKit basic workflow', { repeats: 10 }, async () => {
+			await run2BrowserTest(app, 'livekit', 'custom-emulated', 10);
+		});
+	});
+});
