+++ config/config-all.yaml
@@ -40,21 +40,31 @@ testcases:
     showBrowserVideoElements: true
     description: "Basic 2-user test"
 
-  # browser: multi-emulated: instead of creating one participant at a time, each
-  # room is filled by launching one or more `lk load-test` processes (LiveKit
-  # CLI), each simulating many publishers/subscribers with native simulcast.
-  # Requires a LiveKit platform (platform.url pointing at a LiveKit server).
-  # See README.md "LOADTEST mode" section for what config options do NOT
-  # apply with this browser value (custom media, frameRate, exact resolution, etc.).
+  # browser: emulated (the default): instead of creating one participant at a
+  # time, each room is filled by launching one or more `lk load-test` processes
+  # (LiveKit CLI), each simulating many publishers/subscribers with native
+  # simulcast. Requires a LiveKit platform (platform.url pointing at a LiveKit
+  # server). See README.md "LOADTEST mode" section for what config options do
+  # NOT apply with this browser value (custom media, frameRate, exact resolution, etc.).
   - topology: ONE_SESSION_NXM
     participants:
       - "50:20"
     sessions: 1
     resolution: 1280x720
-    browser: multi-emulated
+    browser: emulated
     videoCodec: h264
     simulcast: true
     description: "LOADTEST mode: 50 publishers + 20 subscribers via lk load-test"
+
+  # browser: custom-emulated: launches one lightweight process per participant
+  # (no real browser), giving full control over per-participant behavior at
+  # the cost of higher resource usage than the default `emulated` mode.
+  - topology: N:N
+    participants:
+      - "10"
+    sessions: 1
+    browser: custom-emulated
+    description: "Custom-emulated participants, one process each"
 
 # ---------- Workers ----------
 workers:
