+++ config/config.yaml
@@ -5,7 +5,7 @@
 # Quick start configuration for testing with a local worker. See the first quick start example in the README for more details.
 
 platform:
-  url: https://your-openvidu-url.io
+  url: https://172-31-224-178.openvidu-local.dev:7443
   apiKey: devkey
   apiSecret: secret
